pong/service/requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flower==2.0.1
gitdb==4.0.11
GitPython==3.1.41
humanize==4.11.0
idna==3.10
kombu==5.4.2
oauthlib==3.2.2
pika==1.3.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2==2.9.9
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
smmap==5.0.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
